requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0
requests==2.32.3
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
